checkBalance: shouldHave(exactText) changed to shouldHave(text)

+++ src/test/java/ru/netology/utils/Card.java
@@ -8,10 +8,10 @@ import lombok.Data;
 public class Card {
     private String id;
     private String number;
-    private double balance;
+    private int balance;
     private String lastSymbols;
 
-    public static void transferMoneyCardToCard(Card cardFrom, Card cardTo, double amount) {
+    public static void transferMoneyCardToCard(Card cardFrom, Card cardTo, int amount) {
         cardFrom.setBalance(cardFrom.getBalance() - amount);
         cardTo.setBalance(cardTo.getBalance() + amount);
     }
